change directory name, split rule files
