+++ contrib/react/react_html_parser.yaml
@@ -1,16 +1,16 @@
 rules:
-  - id: react_html_parser
-    metadata:
-      cwe: "CWE-79: Improper Neutralization of Input During Web Page Generation ('Cross-site Scripting')"
-      owasp: "A7: Cross-Site Scripting (XSS)"
-    patterns:
-      - pattern-either:
-          - pattern: |
-              ReactHtmlParser(...)
-      - pattern-not-inside: |
-          import ReactHtmlParser from 'react-html-parser' ;
-    message: |
-      Untrusted user input in rendering reactHTMLParser can lead to an XSS
-    severity: ERROR
-    languages:
-      - javascript
+- id: react_html_parser
+  metadata:
+    cwe: "CWE-79: Improper Neutralization of Input During Web Page Generation ('Cross-site Scripting')"
+    owasp: 'A7: Cross-Site Scripting (XSS)'
+  patterns:
+  - pattern-either:
+    - pattern: |
+        ReactHtmlParser(...)
+  - pattern-not-inside: |
+      import ReactHtmlParser from 'react-html-parser' ;
+  message: |
+    Untrusted user input in rendering reactHTMLParser can lead to an XSS
+  severity: ERROR
+  languages:
+  - javascript
