+++ cs/lang/security/insecure_deserialization/binary_formatter.cs
@@ -1,0 +1,23 @@
+using System.Runtime.Serialization.Formatters.Binary;
+
+namespace InsecureDeserialization
+{
+    public class InsecureBinaryFormatterDeserialization
+    {
+        public void BinaryFormatterDeserialization(string json)
+        {
+            try
+            {
+                BinaryFormatter binaryFormatter = new BinaryFormatter();
+
+                MemoryStream memoryStream = new MemoryStream(Encoding.UTF8.GetBytes(json));
+                binaryFormatter.Deserialize(memoryStream);
+                memoryStream.Close();
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
+    }
+}
