+++ cs/lang/security/insecure_deserialization/binary_formatter.yaml
@@ -1,0 +1,18 @@
+rules:
+- id: insecure-binaryformatter-deserialization
+  severity: WARNING
+  languages:
+  - C#
+  metadata:
+    cwe: 'CWE-502: Deserialization of Untrusted Data'
+    owasp: 'A8: Insecure Deserialization'
+    references:
+    - https://docs.microsoft.com/en-us/dotnet/standard/serialization/binaryformatter-security-guide
+  message: |
+    The BinaryFormatter type is dangerous and is not recommended for data processing. Applications should stop using BinaryFormatter as soon as possible, even if they believe the data they're processing to be trustworthy. BinaryFormatter is insecure and can't be made secure
+  patterns:
+  - pattern-inside: |
+      using System.Runtime.Serialization.Formatters.Binary;
+      ...
+  - pattern: |
+      new BinaryFormatter();
