+++ cs/lang/security/insecure_deserialization/fast_json.cs
@@ -1,0 +1,19 @@
+using fastJSON;
+
+namespace InsecureDeserialization
+{
+    public class InsecureFastJSONDeserialization
+    {
+        public void FastJSONDeserialization(string json)
+        {
+            try
+            {
+                var obj = JSON.ToObject(json, new JSONParameters { BadListTypeChecking = false });
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
+    }
+}
