+++ cs/lang/security/insecure_deserialization/fast_json.yaml
@@ -1,0 +1,21 @@
+rules:
+- id: insecure-fastjson-deserialization
+  severity: WARNING
+  languages:
+  - C#
+  metadata:
+    cwe: 'CWE-502: Deserialization of Untrusted Data'
+    owasp: 'A8: Insecure Deserialization'
+    references:
+    - https://github.com/mgholam/fastJSON#security-warning-update
+  message: |
+    $type extension has the potential to be unsafe, so use it with common sense and known json sources and not public facing ones to be safe
+  patterns:
+    - pattern-inside: |
+        using fastJSON;
+        ...
+    - pattern: |
+        new JSONParameters
+        {
+          BadListTypeChecking = false
+        }
