+++ cs/lang/security/insecure_deserialization/newtonsoft.cs
@@ -1,0 +1,21 @@
+using Newtonsoft.Json;
+
+namespace InsecureDeserialization
+{
+    public class InsecureNewtonsoftDeserialization
+    {
+        public void NewtonsoftDeserialization(string json)
+        {
+            try
+            {
+                JsonConvert.DeserializeObject<object>(json, new JsonSerializerSettings
+                {
+                    TypeNameHandling = TypeNameHandling.All
+                });
+            } catch(Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
+    }
+}
