+++ cs/lang/security/insecure_deserialization/soap_formatter.cs
@@ -1,0 +1,22 @@
+using System.Runtime.Serialization.Formatters.Soap;
+
+namespace InsecureDeserialization
+{
+    public class InsecureSoapFormatterDeserialization
+    {
+        public void SoapFormatterDeserialization(string json)
+        {
+            try
+            {
+                MemoryStream ms = new MemoryStream(Encoding.UTF8.GetBytes(json));
+
+                SoapFormatter soapFormatter = new SoapFormatter();
+                object obj = soapFormatter.Deserialize(ms);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
+    }
+}
