+++ cs/lang/security/insecure_deserialization/soap_formatter.yaml
@@ -1,0 +1,18 @@
+rules:
+- id: insecure-soapformatter-deserialization
+  severity: WARNING
+  languages:
+  - C#
+  metadata:
+    cwe: 'CWE-502: Deserialization of Untrusted Data'
+    owasp: 'A8: Insecure Deserialization'
+    references:
+    - https://docs.microsoft.com/en-us/dotnet/api/system.runtime.serialization.formatters.soap.soapformatter?view=netframework-4.8#remarks
+  message: |
+    The SoapFormatter type is dangerous and is not recommended for data processing. Applications should stop using SoapFormatter as soon as possible, even if they believe the data they're processing to be trustworthy. SoapFormatter is insecure and can't be made secure
+  patterns:
+    - pattern-inside: |
+        using System.Runtime.Serialization.Formatters.Soap;
+        ...
+    - pattern: |
+        new SoapFormatter();
